split up stores
